procuroid-backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
supabase==2.10.0
python-dotenv==1.0.0
elevenlabs==1.3.0
google-genai==0.3.0
requests==2.31.0
gunicorn==21.2.0
twilio==9.2.1